+++ tracing-honeycomb/Cargo.toml
@@ -22,6 +22,7 @@ rand = "0.7"
 chrono = "0.4.9"
 parking_lot = { version = "0.11.1", optional = true }
 uuid = { version = "0.8.1", features = ["v4"] }
+sha-1 = "0.9.4"
 
 [dev-dependencies]
 tracing-attributes = "0.1.5"
