+++ tracing-honeycomb/src/lib.rs
@@ -11,14 +11,21 @@
 //!
 //! As a tracing layer, `TelemetryLayer` can be composed with other layers to provide stdout logging, filtering, etc.
 
+use rand::{self, Rng};
+
 mod honeycomb;
+mod span_id;
+mod trace_id;
 mod visitor;
 
-pub use crate::honeycomb::{HoneycombTelemetry, SpanId, TraceId};
-pub use crate::visitor::HoneycombVisitor;
-use rand::{self, Rng};
+pub use honeycomb::HoneycombTelemetry;
+pub use span_id::SpanId;
+pub use trace_id::TraceId;
 #[doc(no_inline)]
 pub use tracing_distributed::{TelemetryLayer, TraceCtxError};
+pub use visitor::HoneycombVisitor;
+
+pub(crate) mod deterministic_sampler;
 
 /// Register the current span as the local root of a distributed trace.
 ///
@@ -92,7 +99,7 @@ pub fn new_honeycomb_telemetry_layer(
 pub fn new_honeycomb_telemetry_layer_with_trace_sampling(
     service_name: &'static str,
     honeycomb_config: libhoney::Config,
-    sample_rate: u128,
+    sample_rate: u32,
 ) -> TelemetryLayer<HoneycombTelemetry, SpanId, TraceId> {
     let instance_id: u64 = rand::thread_rng().gen();
     TelemetryLayer::new(
